+++ Model/GeometricDistribution.cs
@@ -5,22 +5,22 @@ namespace RandomNumberGenerationAndModeling.Model
 {
     public class GeometricDistribution : DiscreteDistribution
     {
-        public float SuccessProbability { get; set; }
-        public override float MathExpectation => 1 / SuccessProbability;
-        public override float Variance => (1 - SuccessProbability) / SuccessProbability * SuccessProbability;
-        public override float StandardDeviation => (float) Math.Sqrt(Variance);
+        public double SuccessProbability { get; set; }
+        public override double MathExpectation => 1 / SuccessProbability;
+        public override double Variance => (1 - SuccessProbability) / SuccessProbability * SuccessProbability;
+        public override double StandardDeviation => Math.Sqrt(Variance);
 
-        public GeometricDistribution(int length, float successProbability) : base(length)
+        public GeometricDistribution(int length, double successProbability) : base(length)
         {
             SuccessProbability = successProbability;
         }
 
-        public override float GetValue(float x)
+        public override double GetValue(double x)
         {
-            if (Length == (int) x)
-                return 1 - (float)Math.Pow(1 - SuccessProbability, (int)x - 1) * SuccessProbability;
+            if (Length == x)
+                return 1 - Math.Pow(1 - SuccessProbability, x - 1) * SuccessProbability;
 
-            return (float) Math.Pow(1 - SuccessProbability, (int) x - 1) * SuccessProbability;
+            return Math.Pow(1 - SuccessProbability, x - 1) * SuccessProbability;
         }
 
         public override string ToString()
