+++ ViewModel/CltSamplerDialogViewModel.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Windows;
 using RandomNumberGenerationAndModeling.Model;
 
@@ -7,8 +6,8 @@ namespace RandomNumberGenerationAndModeling.ViewModel
 {
     public class CltSamplerDialogViewModel: GeneratorDialogViewModel<CltSampler>
     {
-        public float Shift { get; set; }
-        public float Scale { get; set; }
+        public double Shift { get; set; }
+        public double Scale { get; set; }
         public int PopulationSize { get; set; }
 
         public CltSamplerDialogViewModel(CltSampler cltSampler, Window dialogView) : base(cltSampler, dialogView)
