+++ ViewModel/CongruetialGeneratorDialogViewModel.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Windows;
 using RandomNumberGenerationAndModeling.Model;
 
@@ -7,10 +6,10 @@ namespace RandomNumberGenerationAndModeling.ViewModel
 {
     public class CongruetialGeneratorDialogViewModel: GeneratorDialogViewModel<CongruentialGenerator>
     {
-        public long Modulus { get; set; }
-        public long Seed { get; set; }
-        public long Multiplier { get; set; }
-        public long Increment { get; set; }
+        public double Modulus { get; set; }
+        public double Seed { get; set; }
+        public double Multiplier { get; set; }
+        public double Increment { get; set; }
 
         public CongruetialGeneratorDialogViewModel(CongruentialGenerator congruentialGenerator, Window dialogView) 
             : base(congruentialGenerator, dialogView)
