+++ ViewModel/GeometricDistributionDialogViewModel.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Windows;
 using RandomNumberGenerationAndModeling.Model;
 
@@ -8,7 +7,7 @@ namespace RandomNumberGenerationAndModeling.ViewModel
     public class GeometricDistributionDialogViewModel : DistributionDialogViewModel<GeometricDistribution>
     {
         public int Length { get; set; }
-        public float SuccessProbability { get; set; }
+        public double SuccessProbability { get; set; }
 
         public GeometricDistributionDialogViewModel(GeometricDistribution geometricDistribution, Window dialogView) 
             : base(geometricDistribution, dialogView)
