+++ ViewModel/InverseRayleighFunctionDialogViewModel.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Windows;
 using RandomNumberGenerationAndModeling.Model;
 
@@ -7,7 +6,7 @@ namespace RandomNumberGenerationAndModeling.ViewModel
 {
     public class InverseRayleighFunctionDialogViewModel : DistributionDialogViewModel<InverseRayleighFunction>
     {
-        public float VariableS { get; set; }
+        public double VariableS { get; set; }
 
         public InverseRayleighFunctionDialogViewModel(InverseRayleighFunction inverseRayleighFunction, Window dialogView)
             : base(inverseRayleighFunction, dialogView)
